fix: Fixed the requested changes
