+++ client/src/components/common/FadeIn.tsx
@@ -1,0 +1,47 @@
+import React, { useEffect, useRef, useState } from "react";
+
+const FadeIn = ({
+  children,
+  delay = 0,
+}: {
+  children: React.ReactNode;
+  delay?: number;
+}) => {
+  const [isVisible, setIsVisible] = useState(false);
+  const domRef = useRef<HTMLDivElement>(null);
+
+  useEffect(() => {
+    const observer = new IntersectionObserver(
+      (entries) => {
+        if (entries[0].isIntersecting) {
+          setTimeout(() => {
+            setIsVisible(true);
+          }, delay);
+          if (domRef.current) observer.unobserve(domRef.current);
+        }
+      },
+      { threshold: 0.15 }
+    );
+
+    if (domRef.current) {
+      observer.observe(domRef.current);
+    }
+
+    return () => {
+      if (domRef.current) observer.unobserve(domRef.current);
+    };
+  }, [delay]);
+
+  return (
+    <div
+      ref={domRef}
+      className={`transition-opacity duration-1000 ${
+        isVisible ? "opacity-100" : "opacity-0"
+      }`}
+    >
+      {children}
+    </div>
+  );
+};
+
+export default FadeIn;
