+++ client/src/components/common/Layout.tsx
@@ -4,7 +4,6 @@ const Layout = () => {
       <nav className="bg-black shadow-md">
         <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
           <div className="flex h-16 items-center justify-between">
-            {/* Logo */}
             <div className="flex-shrink-0">
               <a className="text-3xl font-bold text-gray-300" href="/">
                 Marai
